requirements.txt
streamlit==1.28.0
ultralytics==8.4.15
opencv-python-headless==4.9.0.80
numpy==1.26.4
pillow==10.4.0
